requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
psutil>=5.9.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyautogui>=0.9.54
customtkinter>=5.2.0
numpy>=1.24.0
requests>=2.31.0
google-generativeai>=0.3.0